Revert "Step 6a: Initial refactor of batched server"
This reverts commit 33b108df8b7c67ced9fde5d9d67e7e455d1424ca.
